Renaming Project and checkstyle fixes
