+++ fabric-enchanting-api-v1/build.gradle
@@ -1,0 +1,6 @@
+archivesBaseName = "fabric-enchanting-api-v1"
+version = getSubprojectVersion(project, "1.0.0")
+
+dependencies {
+	compile project(path: ':fabric-api-base', configuration: 'dev')
+}
