+++ fabric-enchanting-api-v1/src/main/java/net/fabricmc/fabric/api/enchanting/v1/EnchantmentTargetRegistry.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.enchantment.v1;
+package net.fabricmc.fabric.api.enchanting.v1;
 
-import net.fabricmc.fabric.impl.enchantment.EnchantmentTargetRegistryImpl;
+import net.fabricmc.fabric.impl.enchanting.EnchantmentTargetRegistryImpl;
 
 import java.util.Iterator;
