+++ fabric-enchanting-api-v1/src/main/java/net/fabricmc/fabric/mixin/enchanting/MixinEnchantingTableContainer.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.enchantment;
+package net.fabricmc.fabric.mixin.enchanting;
 
-import net.fabricmc.fabric.api.enchantment.v1.EnchantingPowerProvider;
+import net.fabricmc.fabric.api.enchanting.v1.EnchantingPowerProvider;
 import net.minecraft.block.BlockState;
 import net.minecraft.container.EnchantingTableContainer;
 import net.minecraft.item.ItemStack;
