+++ fabric-enchanting-api-v1/src/main/resources/assets/fabric-enchanting-api-v1/icon.png
