+++ fabric-enchanting-api-v1/src/main/resources/fabric.mod.json
@@ -1,11 +1,11 @@
 {
   "schemaVersion": 1,
-  "id": "fabric-enchantments-v1",
-  "name": "Fabric Enchantments (v1)",
+  "id": "fabric-enchanting-api-v1",
+  "name": "Fabric Enchanting API (v1)",
   "version": "${version}",
   "environment": "*",
   "license": "Apache-2.0",
-  "icon": "assets/fabric-enchantments-v1/icon.png",
+  "icon": "assets/fabric-enchanting-api-v1/icon.png",
   "contact": {
     "homepage": "https://fabricmc.net",
     "irc": "irc://irc.esper.net:6667/fabric",
@@ -20,11 +20,6 @@
   },
   "description": "Adds enchantment and enchanting related hooks.",
   "mixins": [
-    "fabric-enchantments-v1.mixins.json"
-  ],
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.EnchantmentMod"
-    ]
-  }
+    "fabric-enchanting-api-v1.mixins.json"
+  ]
 }
